test housekeeping: fix pre-existing clippy/build errors for --all-targets
`cargo clippy --all-targets -- -D warnings` was broken on this branch
by three pre-existing test-only errors that the standard
`cargo clippy --all` skips (because it doesn't build test targets by
default):

1. `crates/akita-scheme/src/tests.rs:491` — `QuadraticEquation::new_prover`
   call passed 14 args; the current signature takes 13. The drift
   removed the `&batch_setup.expanded` parameter; drop it from the
   test call site to realign.

2. `crates/akita-prover/src/protocol/ring_switch.rs:1316` — redundant
   `centered[i] as i64` cast (`centered` is already `i64`).

3. `crates/akita-algebra/src/ring/ntt_matvec.rs:15-17` — three
   `#[cfg(test)]` imports (`select_crt_ntt_params`, `ProtocolCrtNttParams`,
   `AkitaError`) became unused after the CRT+NTT dispatch move in
   `831ccfc` (the test module that used them lives in
   `crt_ntt_cache.rs` now).

No behaviour change. Enables full `--all-targets` clippy gate in CI.

+++ crates/akita-prover/src/protocol/ring_switch.rs
@@ -1312,8 +1312,7 @@ mod tests {
         const D: usize = 128;
 
         let centered = from_fn(|i| i64::from(((37 * i as i32 + 11) % 95) - 47));
-        let ring =
-            CyclotomicRing::<F, D>::from_coefficients(from_fn(|i| F::from_i64(centered[i] as i64)));
+        let ring = CyclotomicRing::<F, D>::from_coefficients(from_fn(|i| F::from_i64(centered[i])));
 
         for (num_digits, log_basis) in [
             (7usize, 3u32),
